+++ CHANGES.md
@@ -17,6 +17,7 @@
 * fixed `Interval.negative_bias()` return value;
 * extended GitHub Actions workflow for feature branches, and Python 3.14;
 * completed README **Dependencies**, **Examples**, and **Related projects** sections;
+* moved interval demonstration scripts from **tests/** to **examples/**;
 
 
 ## 0.0.2 - 24th August 2025